+++ gDRplots/NEWS.md
@@ -1,3 +1,6 @@
+## gDRplots 0.0.99 - 2026-01-19
+* update `pheatmap_qc` (change standardized concentrations into unique strings)
+
 ## gDRplots 0.0.98 - 2026-01-02
 * update x-axis label for `plot_volcano_assoc_panel`